Remove async from LoraDevice::new().
